+++ test/PHPUnit/HeliosUtilitiesTestTrait.php
@@ -3,8 +3,8 @@
 trait HeliosUtilitiesTestTrait {
 
 	protected function createTransaction(){
-		$sql = "INSERT INTO helios_transactions(user_id,authority_id,last_status_id,filename) VALUES (?,?,?,?) returning ID;";
-		return $this->getSQLQuery()->queryOne($sql,1,1,4,"toto.txt");
+		$sql = "INSERT INTO helios_transactions(user_id,authority_id,last_status_id,filename,sha1) VALUES (?,?,?,?,?) returning ID;";
+		return $this->getSQLQuery()->queryOne($sql,1,1,4,"toto.txt","ab3321d34d3fb32b52332befa534c9854fff677b");
 	}
 
 	/**